// PROJECT.UI/src/app/models/AD/T_AD_SYSTEM_CONFIG.model.ts
export interface T_AD_SYSTEM_CONFIG {
    PKID: string;
    SAP_HOST: string;
    SAP_CLIENT: string;
    SAP_NUMBER: string;
    SAP_USER_NAME: string;
    SAP_PASSWORD: string;
    SAP_TIME_DIFF: number;
    TGBX_URL: string;
    TGBX_USER_NAME: string;
    TGBX_PASSWORD: string;
    TGBX_TIME_DIFF: number;
    TGBX_API_URL: string;
    TGBX_API_USER_NAME: string;
    TGBX_API_PASSWORD: string;
    SMO_API_USER_NAME: string;
    SMO_API_PASSWORD: string;
    SMO_API_LINK_SMO: string;
    SMS_WEBSERVICE: string;
    SMS_APP: string;
    SMS_PASSWORD: string;
    SMS_BRAND_NAME: string;
    MAIL_SMTPHOST: string;
    MAIL_USER: string;
    MAIL_PASSWORD: string;
    MAIL_PORT: number;
    JOB_STATUS_SAP: number;
    JOB_STATUS_TGBX: number;
    JOB_SEND_EMAIL: number;
    JOB_SEND_SMS: number;
    XTTD_TIME_DIFF: number;
    LAST_SYN_VEHICLE: Date;
    EGAS_WEBSERVICE: string;
    EGAS_USER_NAME: string;
    EGAS_PASSWORD: string;
    SMO_KV2_API_LINK: string;
    SMO_KV2_API_USER_NAME: string;
    SMO_KV2_API_PASSWORD: string;
    CREATE_BY: string;
    CREATE_DATE: Date;
    UPDATE_BY: string;
    UPDATE_DATE: Date;
}